requirements.txt
autopep8==1.7.0
Flask==2.2.2
numpy==1.21.5
python-crontab==2.6.0
psutil==5.9.2
mysql-connector-python
